+++ src/CashFlow.Application/CommandHandlers/CreateTransactionCommandHandler.cs
@@ -8,7 +8,7 @@ using CashFlow.Database.Context;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
-namespace CashFlow.Application.Handlers;
+namespace CashFlow.Application.CommandHandlers;
 
 internal sealed class CreateTransactionCommandHandler(ApplicationDbContext dbContext) : IRequestHandler<CreateTransactionCommand, TransactionDto>
 {
