+++ src/CashFlow.Application/QueryHandlers/GetCategoryQueryHandler.cs
@@ -8,7 +8,7 @@ using CashFlow.Database.Context;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
-namespace CashFlow.Application.Handlers;
+namespace CashFlow.Application.QueryHandlers;
 
 internal sealed class GetCategoryQueryHandler(ApplicationDbContext dbContext) : IRequestHandler<GetCategoryQuery, CategoryDto>
 {
