+++ src/CashFlow.Application/QueryHandlers/GetTransactionQueryHandler.cs
@@ -8,7 +8,7 @@ using CashFlow.Database.Context;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
-namespace CashFlow.Application.Handlers;
+namespace CashFlow.Application.QueryHandlers;
 
 internal sealed class GetTransactionQueryHandler(ApplicationDbContext dbContext) : IRequestHandler<GetTransactionQuery, TransactionDto>
 {
